02_pdfAssistant/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn 
openai
groq # groq has hosted open source libraries
sqlalchemy
pgvector
pypdf
psycopg